+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Comment list functionality with `<leader>rl` keymap and `:CodeReviewList` command
+
 ### Fixed
 - Return to normal mode after submitting comment from insert mode